Struct calloop_wayland_source::WaylandSource

pub struct WaylandSource<D> { /* private fields */ }

An adapter to insert an EventQueue into a calloop EventLoop.

This type implements EventSource which generates an event whenever events on the event queue need to be dispatched. The event queue available in the callback calloop registers may be used to dispatch pending events using EventQueue::dispatch_pending.

WaylandSource::insert can be used to insert this source into an event loop and automatically dispatch pending events on the event queue.

Implementations

impl<D> WaylandSource<D>

pub fn new(connection: Connection, queue: EventQueue<D>) -> WaylandSource<D>

Wrap an EventQueue as a WaylandSource.

queue must be from the connection Connection. This is not a safety invariant, but not following this may cause freezes or hangs

pub fn queue(&mut self) -> &mut EventQueue<D>

Access the underlying event queue

Note that you should not replace this queue with a queue from a different Connection, as that may cause freezes or other hangs.

pub fn connection(&self) -> &Connection

Access the connection to the Wayland server

pub fn insert( self, handle: LoopHandle<'_, D>, ) -> Result<RegistrationToken, InsertError<Self>>
where D: 'static,

Insert this source into the given event loop.

This adapter will pass the event loop’s shared data as the D type for the event loop.